chore: add sub component structure
