+++ frontend/src/components/Chat/components/ChatHeader.tsx
@@ -1,14 +1,14 @@
-import Box from "@mui/material/Box";
-import MatrixStatus from "../MatrixStatus";
-import Stack from "@mui/material/Stack";
-import Typography from "@mui/material/Typography";
+import Box from '@mui/material/Box';
+import MatrixStatus from '../../MatrixStatus';
+import Stack from '@mui/material/Stack';
+import Typography from '@mui/material/Typography';
 
 export const ChatHeader: React.FC = () => {
   return (
     <Box
       sx={{
         borderBottom: 1,
-        borderColor: "divider",
+        borderColor: 'divider',
         flexShrink: 0,
         pb: 1,
         mb: 1,
