+++ frontend/src/components/Chat/components/ChatTypingIndicator.tsx
