+++ frontend/src/main.tsx
@@ -1,22 +1,23 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ThemeProvider, CssBaseline, createTheme } from "@mui/material";
-import { RouterProvider } from "react-router-dom";
-import { router } from "./router";
-import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
-import "./styles/app.css";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { ThemeProvider, CssBaseline } from '@mui/material';
+import { RouterProvider } from 'react-router-dom';
+import { router } from './router';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
+import './styles/app.css';
+import theme from './styles/theme';
 
 const queryClient = new QueryClient();
 
-ReactDOM.createRoot(document.getElementById("root")!).render(
+ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <ThemeProvider theme={createTheme()}>
+    <ThemeProvider theme={theme}>
       <CssBaseline />
       <QueryClientProvider client={queryClient}>
         <RouterProvider router={router} />
         <ReactQueryDevtools initialIsOpen={false} />
       </QueryClientProvider>
     </ThemeProvider>
-  </React.StrictMode>
+  </React.StrictMode>,
 );
