+++ frontend/src/styles/theme.ts
@@ -1,0 +1,5 @@
+import { createTheme } from '@mui/material/styles';
+
+const theme = createTheme({});
+
+export default theme;
